Update titles for better SEO and Social Card Relevance
This adds a "title_override" field that can be used on certain
pages to specify entirely what the title should be for social
and SEO purposes. In addition, the overall title in the config
is changed to "The Expanse Cocktails Project", since that's a
better site title for all but one of the pages. The cocktails
index is now changed to having a title_override of
"The Expanse Cocktails Project", and the root index page is changed
to overriding to the original "Paine x Mactane". The title override
currently also overrides the text displayed in the top left nav.

An additional change to make "The Expanse Cocktails Project" link
to the right place in the nav may also be helpful.